refactor(test): further split test_run_manager_issue to meet LOC limits
Move flow resolution and worktree tests to test_run_manager_worktree.py.
