+++ tests/vibe3/commands/test_run_manager_issue.py
@@ -1,4 +1,4 @@
-"""Tests for `vibe3 run --manager-issue` mode."""
+"""Tests for `vibe3 run --manager-issue` basic and session behavior."""
 
 from pathlib import Path
 from unittest.mock import MagicMock
@@ -42,7 +42,7 @@ def _patch_basic(monkeypatch, backend, github, sqlite=None):
 
 
 class TestRunManagerIssueMode:
-    def test_manager_issue_mode_prints_async_session_and_log(self, monkeypatch) -> None:
+    def test_prints_async_session_and_log(self, monkeypatch) -> None:
         backend = _make_backend()
         github = _make_github()
         sqlite = MagicMock()
@@ -72,9 +72,7 @@ class TestRunManagerIssueMode:
             == "manager-orchestrator"
         )
 
-    def test_manager_issue_mode_does_not_use_run_context_builder(
-        self, monkeypatch
-    ) -> None:
+    def test_does_not_use_run_context_builder(self, monkeypatch) -> None:
         backend = _make_backend()
         github = _make_github()
         sqlite = MagicMock()
@@ -96,9 +94,7 @@ class TestRunManagerIssueMode:
         )
         assert "Manage issue #372" in backend.start_async.call_args.kwargs["task"]
 
-    def test_manager_issue_mode_reports_github_timeout_clearly(
-        self, monkeypatch
-    ) -> None:
+    def test_reports_github_timeout_clearly(self, monkeypatch) -> None:
         github = MagicMock()
         github.view_issue.return_value = "network_error"
 
@@ -109,75 +105,7 @@ class TestRunManagerIssueMode:
         assert result.exit_code != 0
         assert "GitHub read timed out or auth/network is unavailable" in result.stderr
 
-    def test_manager_issue_skips_worktree_when_scene_exists(self, monkeypatch) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-        sqlite.get_flows_by_issue.return_value = [
-            {
-                "branch": "task/issue-372",
-                "flow_status": "active",
-                "manager_session_id": None,
-                "updated_at": "2026-04-03T12:00:00",
-            }
-        ]
-        worktree_manager = MagicMock()
-        worktree_manager.resolve_manager_cwd.return_value = (
-            Path("/Users/jacobcy/src/vibe-center/main/.worktrees/issue-372"),
-            False,
-        )
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(
-            run_module.GitClient,
-            "get_git_common_dir",
-            lambda self: "/Users/jacobcy/src/vibe-center/main/.git",
-        )
-        monkeypatch.setattr(
-            run_module,
-            "WorktreeManager",
-            lambda config, repo_root: worktree_manager,
-        )
-        monkeypatch.setattr(
-            run_module, "load_session_id", lambda role, branch=None: None
-        )
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372", "--worktree"])
-
-        assert result.exit_code == 0
-        options = backend.start_async.call_args.kwargs["options"]
-        assert options.worktree is False
-        assert backend.start_async.call_args.kwargs["cwd"] == Path(
-            "/Users/jacobcy/src/vibe-center/main/.worktrees/issue-372"
-        )
-
-    def test_manager_issue_mode_uses_repo_root_cwd_for_first_worktree_launch(
-        self, monkeypatch
-    ) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(
-            run_module.GitClient,
-            "get_git_common_dir",
-            lambda self: "/Users/jacobcy/src/vibe-center/main/.git",
-        )
-        monkeypatch.setattr(
-            run_module, "load_session_id", lambda role, branch=None: None
-        )
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372", "--worktree"])
-
-        assert result.exit_code == 0
-        assert backend.start_async.call_args.kwargs["cwd"] == Path(
-            "/Users/jacobcy/src/vibe-center/main/.worktrees/issue-372"
-        )
-
-    def test_manager_issue_mode_persists_async_session_id_from_log(
-        self, monkeypatch, tmp_path
-    ) -> None:
+    def test_persists_async_session_id_from_log(self, monkeypatch, tmp_path) -> None:
         log_path = tmp_path / "vibe3-manager-issue-372.async.log"
         log_path.write_text("SESSION_ID: ses_manager372\n")
 
@@ -204,7 +132,7 @@ class TestRunManagerIssueMode:
             == "ses_manager372"
         )
 
-    def test_manager_issue_mode_persists_async_session_id_from_wrapper_log(
+    def test_persists_async_session_id_from_wrapper_log(
         self, monkeypatch, tmp_path
     ) -> None:
         wrapper_log = tmp_path / "codeagent-wrapper-53796.log"
@@ -235,9 +163,7 @@ class TestRunManagerIssueMode:
             == "ses_wrapper372"
         )
 
-    def test_manager_issue_mode_reuses_existing_session_for_launch_cwd(
-        self, monkeypatch
-    ) -> None:
+    def test_reuses_existing_session_for_launch_cwd(self, monkeypatch) -> None:
         backend = _make_backend()
         github = _make_github()
         sqlite = MagicMock()
@@ -253,156 +179,7 @@ class TestRunManagerIssueMode:
         assert backend.start_async.call_args.kwargs["session_id"] == "ses_existing"
         assert backend.start_async.call_args.kwargs["cwd"] == Path.cwd()
 
-    def test_manager_issue_mode_prefers_target_issue_flow_branch(
-        self, monkeypatch
-    ) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-        sqlite.get_flows_by_issue.return_value = [
-            {
-                "branch": "task/issue-372",
-                "manager_session_id": "ses_target",
-                "updated_at": "2026-04-03T12:00:00",
-            }
-        ]
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(
-            run_module, "load_session_id", lambda role, branch=None: "ses_target"
-        )
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372"])
-
-        assert result.exit_code == 0
-        assert backend.start_async.call_args.kwargs["session_id"] == "ses_target"
-        sqlite.add_event.assert_called_once()
-        assert sqlite.add_event.call_args.args[0] == "task/issue-372"
-
-    def test_manager_issue_mode_prefers_active_flow_over_aborted_session_flow(
-        self, monkeypatch
-    ) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-        sqlite.get_flows_by_issue.return_value = [
-            {
-                "branch": "do/20260403-aaf82f",
-                "flow_status": "aborted",
-                "manager_session_id": "ses_old",
-                "updated_at": "2026-04-03T12:59:00",
-            },
-            {
-                "branch": "task/issue-372",
-                "flow_status": "active",
-                "manager_session_id": None,
-                "updated_at": "2026-04-01T19:50:00",
-            },
-        ]
-
-        def load_session(role: str, branch: str | None = None) -> str | None:
-            if branch == "task/issue-372":
-                return None
-            if branch == "do/20260403-aaf82f":
-                return "ses_old"
-            return None
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(run_module, "load_session_id", load_session)
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372"])
-
-        assert result.exit_code == 0
-        assert backend.start_async.call_args.kwargs["session_id"] is None
-        assert sqlite.add_event.call_args.args[0] == "task/issue-372"
-
-    def test_manager_issue_mode_uses_target_scene_cwd_for_foreign_flow(
-        self, monkeypatch
-    ) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-        sqlite.get_flows_by_issue.return_value = [
-            {
-                "branch": "task/issue-372",
-                "flow_status": "active",
-                "manager_session_id": None,
-                "updated_at": "2026-04-03T12:00:00",
-            }
-        ]
-        worktree_manager = MagicMock()
-        worktree_manager.resolve_manager_cwd.return_value = (
-            Path("/Users/jacobcy/src/vibe-center/main/.worktrees/issue-372"),
-            False,
-        )
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(
-            run_module.GitClient,
-            "get_git_common_dir",
-            lambda self: "/Users/jacobcy/src/vibe-center/main/.git",
-        )
-        monkeypatch.setattr(
-            run_module, "load_session_id", lambda role, branch=None: None
-        )
-        monkeypatch.setattr(
-            run_module,
-            "WorktreeManager",
-            lambda config, repo_root: worktree_manager,
-        )
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372"])
-
-        assert result.exit_code == 0
-        assert backend.start_async.call_args.kwargs["cwd"] == Path(
-            "/Users/jacobcy/src/vibe-center/main/.worktrees/issue-372"
-        )
-
-    def test_manager_issue_mode_uses_repo_root_for_first_worktree_on_foreign_flow(
-        self, monkeypatch
-    ) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-        sqlite.get_flows_by_issue.return_value = [
-            {
-                "branch": "task/issue-372",
-                "flow_status": "active",
-                "manager_session_id": None,
-                "updated_at": "2026-04-03T12:00:00",
-            }
-        ]
-        worktree_manager = MagicMock()
-        worktree_manager.resolve_manager_cwd.return_value = (None, False)
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(
-            run_module.GitClient,
-            "get_git_common_dir",
-            lambda self: "/Users/jacobcy/src/vibe-center/main/.git",
-        )
-        monkeypatch.setattr(
-            run_module, "load_session_id", lambda role, branch=None: None
-        )
-        monkeypatch.setattr(
-            run_module,
-            "WorktreeManager",
-            lambda config, repo_root: worktree_manager,
-        )
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372", "--worktree"])
-
-        assert result.exit_code == 0
-        assert backend.start_async.call_args.kwargs["cwd"] == Path(
-            "/Users/jacobcy/src/vibe-center/main"
-        )
-        options = backend.start_async.call_args.kwargs["options"]
-        assert options.worktree is True
-        worktree_manager.resolve_manager_cwd.assert_called_once()
-
-    def test_manager_issue_mode_fresh_session_skips_session_resume(
-        self, monkeypatch
-    ) -> None:
+    def test_fresh_session_skips_session_resume(self, monkeypatch) -> None:
         backend = _make_backend()
         github = MagicMock()
         github.view_issue.return_value = {
@@ -424,22 +201,3 @@ class TestRunManagerIssueMode:
 
         assert result.exit_code == 0
         assert backend.start_async.call_args.kwargs["session_id"] is None
-
-    def test_manager_issue_mode_uses_canonical_target_branch_without_target_flow(
-        self, monkeypatch
-    ) -> None:
-        backend = _make_backend()
-        github = _make_github()
-        sqlite = MagicMock()
-        sqlite.get_flows_by_issue.return_value = []
-
-        _patch_basic(monkeypatch, backend, github, sqlite)
-        monkeypatch.setattr(
-            run_module, "load_session_id", lambda role, branch=None: None
-        )
-
-        result = runner.invoke(cli_app, ["run", "--manager-issue", "372"])
-
-        assert result.exit_code == 0
-        sqlite.add_event.assert_called_once()
-        assert sqlite.add_event.call_args.args[0] == "task/issue-372"
